# backend/swagger.yaml
swagger: "2.0"
info:
  description: "Implement UBER like service for freight trucks, in REST style, using MongoDB as database. This service should help regular people to deliver their stuff and help drivers to find loads and earn some money. Application contains 2 roles, driver and shipper."
  version: "1.0.0"
  title: "API service in REST style"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "Dmytro_Ostretsov@epam.com"
host: "localhost:8080"
basePath: "/"
...
schemes:
  - "http"
paths:
  /api/users/me:
    get:
      tags:
        - "User"
      summary: "Get user's profile info"
      description: "User can request only his own profile info"
      operationId: "getProfileInfo"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        "200":
          description: "Success"
          schema:
            type: object
            properties:
              user:
                $ref: "#/definitions/User"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Error"
      security:
        - jwt_token: []
    delete:
      tags:
        - "User"
      summary: "Delete user's profile"
      description: "User can delete only his own profile info"
      operationId: "deleteProfile"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        "200":
          description: "Success"
          schema:
            type: "object"
            properties:
              message:
                type: string
                example: "Profile deleted successfully"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Error"
      security:
        - jwt_token: []
...
  /api/auth/forgot_password:
    post:
      tags:
        - "Auth"
      summary: "Forgot password"
      description: "(Optional) Forgot password option"
      operationId: "forgotPassword"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "credentials"
          schema:
            properties:
              email:
                type: "string"
                example: "kyrylo@gmail.com"
      responses:
        "200":
          description: "Success"
          schema:
            type: "object"
            properties:
              message:
                type: string
                example: "New password sent to your email address"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Error"
...
securityDefinitions:
  jwt_token:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  Error:
    type: "object"
    properties:
      message:
        type: "string"
        description: "Error message"
...
  User:
    type: "object"
    properties:
      _id:
        type: "string"
        example: "5099803df3f4948bd2f98391"
      email:
        type: "string"
        example: "kyrylo@gmail.com"
      created_date:
        type: "string"
        format: "date"
        example: "2020-10-28T08:03:19.814Z"
...